redleft auto paths updated
