+++ MeepMeepPathing/src/main/java/com/example/meepmeeppathing/BlueLeft/BlueLeftAuto.java
@@ -1,4 +1,4 @@
-package com.example.meepmeeppathing;
+package com.example.meepmeeppathing.BlueLeft;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.geometry.Vector2d;
